Simple preprocessor for conditional wgsl compilation
Currently allows things like

```
\#ifdef SOME_DEFINE
...
\#else SOME_DEFINE
...
\#endif SOME_DEFINE
```

else is optional, but the name must appear in all three statements (to
  simplify parsing).

This might not be the final design, we should follow the progress of https://github.com/wgsl-tooling-wg/wesl-spec
and use their syntax (e.g. 'import ngsolve_webgpu::clipping', or '@if' instead of #if) and tooling.

This might also be benficial for development (IDE support), see https://github.com/wgsl-analyzer/wgsl-analyzer/issues/182